refactoring(post): changed post POST parameters during creation
